requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
pymysql==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.28
starlette==0.36.3
strawberry-graphql==0.220.0
typing-extensions==4.10.0
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
